src/modules/product_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pymysql
cryptography
